requirements.txt
numpy>=1.20.0
matplotlib>=3.4.0
pyyaml>=5.4.0
cycler>=0.10.0